crib_engine/requirements.txt
flask==3.0.3
requests==2.32.3
python-dotenv==1.0.1
pandas
SQLAlchemy==2.0.36
pyodbc==5.2.0